add simple auth using session (a little hardcode), get api using ajax

+++ Controllers/EventController.cs
@@ -2,7 +2,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Net.Http;
+using System.Net;
+using System.Net.Http.Headers;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using AdminIB.Models;
@@ -21,12 +25,22 @@ namespace AdminIB.Controllers
         // GET: Event
         public async Task<IActionResult> Index()
         {
+            if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             return View(await _context.Event.ToListAsync());
         }
 
         // GET: Event/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             if (id == null)
             {
                 return NotFound();
@@ -42,9 +56,43 @@ namespace AdminIB.Controllers
             return View(@event);
         }
 
+        public async Task<IActionResult> Detail()
+        {
+            IEnumerable<Event> events = null;
+            // // https://localhost:5001/api/requestevent
+            // using (var client = new HttpClient())
+            // {
+            //     client.BaseAddress = new Uri("https://localhost:5001/");
+            //     //HTTP GET
+            //     client.DefaultRequestHeaders.Accept.Clear();
+            //     client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+            //     client.DefaultRequestHeaders.Add("User-Agent", ".NET Foundation Repository Reporter");
+            //     ServicePointManager.ServerCertificateValidationCallback += (sender, certificate, chain, sslPolicyErrors) => certificate.Issuer == "CN=localhost";
+            //     HttpResponseMessage response = await client.GetAsync("api/requestevent");
+            //     if (response.IsSuccessStatusCode)
+            //     {
+            //         events = await response.Content.ReadAsAsync<List<Event>>();
+            //     }
+            //     else //web api sent error response 
+            //     {
+            //         //log response status here..
+
+            //         events = Enumerable.Empty<Event>();
+
+            //         ModelState.AddModelError(string.Empty, "Server error. Please contact administrator.");
+            //     }
+            // }
+            return View(events);
+        }
+
         // GET: Event/Create
         public IActionResult Create()
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             return View();
         }
 
@@ -55,6 +103,11 @@ namespace AdminIB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,NamaEvent,StartDate,EndDate,Description,Penyelenggara,Keterangan")] Event @event)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             if (ModelState.IsValid)
             {
                 _context.Add(@event);
@@ -67,6 +120,11 @@ namespace AdminIB.Controllers
         // GET: Event/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             if (id == null)
             {
                 return NotFound();
@@ -87,6 +145,11 @@ namespace AdminIB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,NamaEvent,StartDate,EndDate,Description,Penyelenggara,Keterangan")] Event @event)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             if (id != @event.Id)
             {
                 return NotFound();
@@ -107,6 +170,7 @@ namespace AdminIB.Controllers
                     }
                     else
                     {
+                    https://localhost:5001/api/requestevent
                         throw;
                     }
                 }
@@ -118,6 +182,11 @@ namespace AdminIB.Controllers
         // GET: Event/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             if (id == null)
             {
                 return NotFound();
@@ -138,6 +207,11 @@ namespace AdminIB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+             if (HttpContext.Session.GetString("Username") == null)
+            {
+                return RedirectToAction("Login", "Account", new { area = "" });
+            }
+            ViewBag.Username = HttpContext.Session.GetString("Username");
             var @event = await _context.Event.FindAsync(id);
             _context.Event.Remove(@event);
             await _context.SaveChangesAsync();
